% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_IH.Parameters.R
\name{.parameters}
\alias{.parameters}
\title{object, must be one of \{"quad", "uni", "exp"\} indicating the distribution}
\usage{
.parameters(
  timePoints,
  tau,
  nTimes,
  response,
  nTree,
  ERT,
  uniformSplit,
  randomSplit,
  splitRule,
  replace,
  nodeSize,
  minEvent,
  tieMethod,
  criticalValue,
  nSamples,
  stratifiedSplit
)
}
\description{
object, must be one of \{"quad", "uni", "exp"\} indicating the distribution
}
