% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IH.dtrSurv.R
\name{IHsurvrf}
\alias{IHsurvrf}
\title{Indefinite Horizon Dynamic Treatment Regime Estimation for Survival Outcomes}
\usage{
IHsurvrf(
  data,
  txName,
  models,
  ...,
  timePoints = "uni",
  nTimes = 200L,
  tau = NULL,
  criticalValue = "mean",
  splitRule = NULL,
  ERT = TRUE,
  uniformSplit = NULL,
  sampleSize = NULL,
  replace = NULL,
  randomSplit = 0.2,
  tieMethod = "random",
  minEvent = 3L,
  nodeSize = 6L,
  nTree = 10L,
  mTry = NULL,
  stratifiedSplit = NULL,
  stageLabel = ".",
  nstrata = 1,
  windowsize = 10
)
}
\arguments{
\item{data}{A data.frame object. The full dataset including patient ID, treatments
received, all stage covariates, observed times, and censoring
indicators.
Can be provided as a matrix object if column headers are included.
Can contain missing data coded as NA, but cannot contain NaN.
If performing analysis with multiple strata, cumulative.time is also required for strata splitting.}

\item{txName}{A character vector object. The treatment variable name for
each decision point. Each element corresponds to the respective decision
point (element 1 = 1st decision; element 2 = 2nd decision, etc.).}

\item{models}{A single formula. Contains a formula defining the
response as a Surv() object and the covariate structure of the model.
Note that this model should not include any terms of order > 1. This
is a common formula to be used across all decision
points.
Example: Surv(visit.length, delta) ~ A + prev.visit.length
In this example, visit.length is the observed time, delta is the censoring indicator, A is treatment.
Any number of covariates can be added in the model.}

\item{...}{Ignored. Present only to require named inputs.}

\item{timePoints}{A character object or a numeric vector object. If a character
object, must be one of \{"quad", "uni", "exp"\} indicating the distribution
from which the time points are to be calculated. For character input,
input 'nTimes' must also be provided. If a numeric vector, the
time points to be used. If 0 is not the first value, it will be
concatenated by the software.}

\item{nTimes}{An integer object. The total number of time points to be
generated and considered. Used in conjunction with input 'timePoints'
when 'timePoints' is a character; ignored otherwise.}

\item{tau}{A numeric object or NULL. The study length. If NULL, the
maximum timePoint is used.}

\item{criticalValue}{A character object. Must be \{"mean"\}, which is the
estimator for the value of a treatment rule which is the mean survival time.}

\item{splitRule}{A character object OR NULL.
Must be \{"mean"\} indicating the test used to determine an optimal split.
Indicates usage of the truncated mean test.}

\item{ERT}{A logical object. If TRUE, the Extremely Randomized Trees
algorithm is used to select the candidate variable.}

\item{uniformSplit}{A logical object. If 'ERT' and 'uniformSplit' are TRUE,
the random cutoff is sampled from a uniform distribution over the range
of available covariate values. If 'ERT' is TRUE and 'uniformSplit' is
FALSE, a case is randomly selected and the cutoff is taken to be the mean
cutoff between it and the next largest covariate value. If 'ERT' is FALSE,
input is ignored.}

\item{sampleSize}{A numeric object, numeric vector object, or NULL.
The fraction (0 < sampleSize <= 1) of the data to be used for each
tree in the forest. If only
one value is given, it is assumed to be the fraction for all decision
points. If a vector is given, the length must be equal to the total
number of decision points and each element corresponds to its respective
decision point. If NULL and 'ERT' is TRUE,
sampleSize defaults to 1.0. If NULL and 'ERT'
is FALSE, sampleSize defaults to 0.632.}

\item{replace}{A logical object or NULL. If TRUE, the sample drawn for each
of the nTree trees may have duplicate records. If FALSE, no individual is
present in the sample for than once. If NULL, 'replace' = !'ERT'.}

\item{randomSplit}{A numeric object. The probability that a random split
will occur. Must be 0 < randomSplit < 1.}

\item{tieMethod}{A character object. Must be one of
\{"first", "random"\}. If multiple splits lead to the same
value, the method by which the tie is broken.}

\item{minEvent}{An integer object. The minimum number of events that must be
present in a node.}

\item{nodeSize}{An integer object. The minimum number of individuals that
must be present in a node.}

\item{nTree}{An integer object. The number of trees to grow.}

\item{mTry}{An integer or integer vector object. The maximum number of
covariates to sample for each split. If a vector, each element
corresponds to its respective decision point.}

\item{stratifiedSplit}{A numeric object. The stratified random split
coefficient. Covariates for which the number of splits (s_i) is less
than s*stratifiedSplit/d are explored preferentially
total number of covariates under consideration).}

\item{stageLabel}{A character object. If using a common formula, the
character used to separate the covariate from the decision point label.
Must be "\emph{".
Example: "}", if the decision points are labeled Y_1, Y_2, Y_2}
}
\value{
An S4 object of class DTRSurvRes containing the key results and
input parameters of the analysis.
}
\description{
Provides methods for estimating multi-stage (indefinite horizon) optimal dynamic treatment
regimes for survival outcomes, specifically truncated mean survival time, with dependent censoring.
}
\details{
----------------- DESCRIPTION -----------------
We use a common formula for all decision points, i.e.,
'models' is a single formula object; therefore, your data must follow a specific
format. Specifically, 'stageLabel' = "_" is a requirement, so covariates must be named as
xxx_1 for the first decision point,
xxx_2 for the second, xxx_3 for the third, etc. The exact structure of the
'xxx' can be generally defined; however, it cannot contain the stageLabel.
For example, if the column names are
(Y_1, Y_2, delta_1, delta_2, A_1, A_2, X_1, X_2)
'models' = Surv(Y,delta) ~ X + A would lead to
Surv(Y_1, delta_1) ~ X_1 + A_1 as the first stage model; and
Surv(Y_2, delta_2) ~ X_2 + A_2
as the second stage.
The censoring indicator column, if using more than 1 strata, must be named "delta",
which will be in the Surv( xx, delta). Otherwise, it can be named in any manner.

Y_k is the length of Stage k so that (Y_1 + Y_2 + ... + Y_K) is the
overall observed failure time, delta_k is the censoring status at Stage k,
d_k = 0 if a subject was censored at Stage k, and 1 if he/she experienced
failure during that stage or moved to Stage k+1.
A_k is the treatment at Stage k, k=1,2,..., K. Note that every quantity
here is stage-wide. In other words, Y_2 is the length of Stage 2
and is not cumulative from the baseline.

When one experienced censoring or failure at Stage k, it should be that
Y_j = 0 for all j > k and instantaneous failure (Y_k < 1e-8) is not allowed;
E.g., when delta_(k-1) = 1 and Y_k = 0, the person is considered dead at
Stage k-1, but when d_(k-1) = 1 and Y_k = 2, the person made it to Stage k
and either experienced failure or censoring (depending on d_k) during Stage k.

Any subject with missing values at Stage k will be ignored.
}
\examples{


dt <- data.frame("Y_1" = sample(1:100,100,TRUE), "Y_2" = sample(1:100,100,TRUE),
                 "D_1" = rbinom(100, 1, 0.9), "D_2" = rbinom(100,1,0.9),
                 "A_1" = rbinom(100, 1, 0.5), "A_2" = rbinom(100,1,0.5),
                 "X_1" = rnorm(100), "X_2" = rnorm(100), subj.id = 1:100)

2 stage analysis with 1 strata
IHsurvrf(data = dt,
        txName = c("A_1", "A_2"),
        stageLabel = "_",
        models = Surv(Y,D)~X+A,
        nstrata = 1)



 5 stage analysis with 2 strata. We note that cumulative.time is required for more than 1 strata.
 We also note that due to the use of more than 1 strata, we need to name the censoring indicator "delta"

dt <- data.frame("Y_1" = sample(1:100,100,TRUE), "Y_2" = sample(1:100,100,TRUE),
                 "Y_3" = sample(1:100,100,TRUE), "Y_4" = sample(1:100,100,TRUE),
                 "Y_5" = sample(1:100,100,TRUE),
                  "delta_1" = rbinom(100, 1, 0.9), "delta_2" = rbinom(100,1,0.9),
                 "delta_3" = rbinom(100,1,0.9), "delta_4" = rbinom(100,1,0.9),
                 "delta_5" = rbinom(100,1,0.9),
                  "A_1" = rbinom(100, 1, 0.5), "A_2" = rbinom(100,1,0.5),
                 "A_3" = rbinom(100,1,0.5), "A_4" = rbinom(100,1,0.5),
                 "A_5" = rbinom(100,1,0.5),
                  "X_1" = rnorm(100), "X_2" = rnorm(100),
                 "X_3" = rnorm(100), "X_4" = rnorm(100),
                 "X_5" = rnorm(100), subj.id = 1:100)

dt$cumulative.time_1 <- 0
dt$cumulative.time_2 <- dt$Y_1
dt$cumulative.time_3 <- dt$Y_1 + dt$Y_2
dt$cumulative.time_4 <- dt$Y_1 + dt$Y_2 + dt$Y_3
dt$cumulative.time_5 <- dt$Y_1 + dt$Y_2 + dt$Y_3 + dt$Y_4

IHsurvrf(data = dt,
        txName = c("A_1", "A_2", "A_3", "A_4", "A_5"),
        stageLabel = "_",
        models = Surv(Y,delta)~X+A,
        nstrata = 2, windowsize = 5)
}
\references{
She, J., Egberg, M., and Kosorok, M.R.
An optimal dynamic treatment regime estimator for indefinite-horizon survival outcomes.
Submitted.
}
\seealso{
\code{\link{predict}} for retrieving the optimal treatment
and/or the optimal survival curves.
}
