% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_IH.DTRSurv.R
\name{predict}
\alias{predict}
\alias{predict,DTRSurv-method}
\title{Method to estimate the value for new data or to retrieve estimated value for
training data}
\usage{
\S4method{predict}{DTRSurv}(object, ..., newdata, findOptimal = TRUE)
}
\arguments{
\item{object}{A DTRSurv object. The object returned by a call to dtrSurv().}

\item{...}{Ignored. Used to require named inputs.}

\item{newdata}{NULL or a data.frame object. If NULL, this method retrieves
the estimated value for the training data. If a data.frame, the
value is estimated based on the data provided.}

\item{findOptimal}{A logical object. If TRUE, the value is estimated for
all treatment options and that leading to the maximum value for each
individual is used to estimate the value.}
}
\value{
a list object containing a matrix of the predicted survival function
(survFunc), the estimated mean survuval (mean), and the estimated
survival probability (if critical value is surv.mean or surv.prob)
}
\description{
Method to estimate the value for new data or to retrieve estimated value for
training data
}
\examples{


dt <- data.frame("Y_1" = sample(1:100,100,TRUE), "Y_2" = sample(101:200,100,TRUE),
                 "D_1" = rbinom(100, 1, 0.9), "D_2" = rbinom(100,1,0.9),
                 "A_1" = rbinom(100, 1, 0.5), "A_2" = rbinom(100,1,0.5),
                 "X_1" = rnorm(100), "X_2" = rnorm(100), subj.id = 1:100)

result <- IHsurvrf(data = dt,
                  txName = c("A_1", "A_2"),
                  models = Surv(Y,D)~X+A, stageLabel = "_",)

tt <- predict(object = result@Forest1)
tt <- predict(object = result@Forest1, findOptimal = FALSE)
}
